+++ di/src/main/java/com/interface21/beans/factory/config/BeanDefinitionMapping.java
@@ -1,26 +1,33 @@
 package com.interface21.beans.factory.config;
 
 import com.interface21.beans.BeanScanner;
-import com.interface21.beans.factory.support.BeanFactoryUtils;
-import com.interface21.context.annotation.Bean;
+import com.interface21.beans.factory.support.BeanDefinitionReader;
+import com.interface21.beans.factory.support.BeanDefinitionRegistry;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class BeanDefinitionMapping {
+public class BeanDefinitionMapping implements BeanDefinitionRegistry {
     private final Map<String, BeanDefinition> beanDefinitionMap = new HashMap<>();
-
+    private final BeanDefinitionReader reader;
     private final BeanScanner scanner;
 
     public BeanDefinitionMapping(final String... basePackages) {
+        this.reader = new ConfigurationClassBeanDefinitionReader(this);
         this.scanner = new BeanScanner((Object) basePackages);
+    }
+
+    @Override
+    public void registerBeanDefinition(final Class<?> clazz, final BeanDefinition beanDefinition) {
+        beanDefinitionMap.put(clazz.getName(), beanDefinition);
     }
 
     public void scanBeanDefinitions() {
         final Set<Class<?>> beanClasses = scanner.scanClassesTypeAnnotatedWith();
         toBeanDefinitionMap(beanClasses);
+        reader.loadBeanDefinitions(beanClasses.toArray(new Class<?>[0]));
     }
 
     private void toBeanDefinitionMap(final Set<Class<?>> beanClasses) {
@@ -28,17 +35,7 @@ public class BeanDefinitionMapping {
     }
 
     private void put(final Class<?> beanClass) {
-        putBeanMethods(beanClass);
-
-        beanDefinitionMap.put(beanClass.getName(), new RootBeanDefinition(beanClass, beanClass.getName()));
-    }
-
-    private void putBeanMethods(final Class<?> beanClass) {
-        BeanFactoryUtils.getBeanMethods(beanClass, Bean.class).forEach(beanMethod -> {
-            final ConfigurationClassBeanDefinition beanDefinition = new ConfigurationClassBeanDefinition(beanMethod, beanClass);
-
-            beanDefinitionMap.put(beanDefinition.getBeanClassName(), beanDefinition);
-        });
+        registerBeanDefinition(beanClass, new RootBeanDefinition(beanClass, beanClass.getName()));
     }
 
     public Set<Class<?>> getBeanClasses() {
